+++ flaskbb/core/search/backends/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""
+flaskbb.core.search.backends
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The concrete search backend implementations. Each implements the
+`SearchBackend` interface from `flaskbb.core.search.base` and is selected
+by name via the `SEARCH_BACKEND` config (see `flaskbb.core.search`).
+
+:copyright: (c) 2014-2026 by the FlaskBB Team.
+:license: BSD, see LICENSE for more details.
+"""
